+++ http-verbs-common/src/main/scala/uk/gov/hmrc/http/HttpClient.scala
@@ -18,16 +18,16 @@ package uk.gov.hmrc.http
 
 import akka.actor.ActorSystem
 import com.typesafe.config.{Config, ConfigFactory}
-import play.api.libs.ws.{DefaultWSProxyServer, WSClient, WSProxyServer}
+import play.api.libs.ws.{WSClient, WSProxyServer}
 import uk.gov.hmrc.http.hooks.HttpHook
 import uk.gov.hmrc.play.http.ws._
 
 trait HttpClient extends HttpGet with HttpPut with HttpPost with HttpDelete with HttpPatch {
-  // TODO require implementations to not break clients e.g. implementations of ProxyHttpClient...
+  // implementations required to not break clients (which won't be using the new functions) e.g. implementations of ProxyHttpClient...
   def withUserAgent(userAgent: String): HttpClient =
-    sys.error("Your implementation of HttpClient does not implement `withUserAgent`. Consider if you can use uk.gov.hmrc.play.bootstrap.http.DefaultHttpClient") // TODO reference to bootstrap :(
+    sys.error("Your implementation of HttpClient does not implement `withUserAgent`. Consider if you can use uk.gov.hmrc.http.bootstrap.http.DefaultHttpClient") // TODO reference to bootstrap :(
 
-  def withProxy(): HttpClient =
+  def withProxy: HttpClient =
     sys.error("Your implementation of HttpClient does not implement `withProxy`. Consider if you can use uk.gov.hmrc.play.bootstrap.http.DefaultHttpClient") // TODO reference to bootstrap :(
 }
 
@@ -50,7 +50,7 @@ class HttpClientImpl (
       actorSystem
     )
 
-  override def withProxy(): HttpClientImpl =
+  override def withProxy: HttpClientImpl =
     new HttpClientImpl(
       configuration,
       hooks,
@@ -58,22 +58,6 @@ class HttpClientImpl (
       actorSystem
     ) {
       override val wsProxyServer: Option[WSProxyServer] =
-        if (configuration.getBoolean("proxy.proxyRequiredForThisEnvironment")) { // keep this check to avoid needing the following configuration for development? rename to `proxy.enabled`?
-          Some(
-            DefaultWSProxyServer(
-              protocol  = Some(configuration.getString("proxy.protocol")),
-              host      = configuration.getString("proxy.host"),
-              port      = configuration.getInt("proxy.port"),
-              principal = getOptionalString("proxy.username"),
-              password  = getOptionalString("proxy.password"),
-              // following exists to be development friendly (necessary with `proxy.proxyRequiredForThisEnvironment`?)
-              nonProxyHosts = Some(Seq("localhost"))
-              //nonProxyHosts = Some(configuration.underlying.getStringList("proxy.nonProxyHosts").asScala.toSeq)
-            )
-          )
-        } else None
-
-      private def getOptionalString(key: String): Option[String] =
-        if (configuration.hasPath(key)) Some(configuration.getString(key)) else None
+        WSProxyConfiguration.buildWsProxyServer(configuration)
     }
 }